Cell down alarm notification [Issue-Id: ODUHIGH-430]

+++ src/o1/ves/SliceMeasurementEvent.cpp
@@ -122,7 +122,7 @@ bool SliceMeasurementEvent::prepareEventFields(const Message* msg)
 
     if(JsonHelper::addNodeToObject(measurementFields, \
                                    "measurementInterval", \
-                                    MEASUREMENT_INTERVAL) == 0)
+                                    (double)MEASUREMENT_INTERVAL) == 0)
     {
         ret = false;
     }
